requirements.txt
pandas>=2.0.0
duckdb>=0.8.0
openpyxl>=3.1.0
xlwings>=0.30.0
pycountry>=22.0.0
matplotlib>=3.7.0
numpy>=1.24.0
pyxlsb>=1.0.10
gdxpds>=1.0.0
seaborn>=0.12.0
h5py>=3.8.0
geopandas>=1.1.0
scikit-learn>=1.7.0
PyYAML>=6.0.0 